Description
This project is a TensorFlow-based Deep Learning classifier for real-time anomaly detection to improve commercial and national security. It becomes even more complicated and sophisticated when dealing with highly populated countries like India. The prototype consists of 4 modules: data organization, video to frame conversion, feature extraction and model evaluation. UCF Crime Dataset is chosen as it contains a variety of video clips based on real-world anomalies to ensure the outcomes are tangible. The entire dataset of videos is transformed into individual frames using the FFmpeg library to get the features from each frame. The feature extraction process is possible using Convolutional Neural Networks, and the data is saved for training. The elements are appended to the frames in a fixed sequence fed to LSTM to train the entire dataset. Added callbacks ensure the model holds the best weights in each epoch to attain more accuracy.